Apply confirm box

+++ src/app/project/containers/responses/responses.component.ts
@@ -3,6 +3,7 @@ import { Router } from '@angular/router'
 import { Store } from '@ngrx/store'
 import { ResponseService } from 'app/project/services/response.service'
 import { EndpointService } from 'app/project/services/endpoint.service'
+import { ConfirmService } from 'app/shared/services/confirm.service'
 import * as database from 'app/core/services/database.service'
 import * as fromProject from 'app/project/reducers'
 
@@ -27,7 +28,8 @@ export class ResponsesComponent implements OnInit {
     private store: Store<any>,
     private router: Router,
     private responseService: ResponseService,
-    private endpointService: EndpointService
+    private endpointService: EndpointService,
+    private confirmService: ConfirmService
   ) {
     this.tabSelector = this.tabAll[0]
     this.store.select(fromProject.getEndpointId)
@@ -111,19 +113,34 @@ export class ResponsesComponent implements OnInit {
   }
 
   deleteResponse (id) {
-    this.responseService.delete(id)
-      .subscribe(res => {
-        if (!res.error) {
-          this.responses = this.responses.filter(response => response.id !== id)
+    this.confirmService.open({
+      message: 'Are you sure you want to delete this endpoint'
+    })
+      .afterClose(val => {
+        if (val) {
+          this.responseService.delete(id)
+            .subscribe(res => {
+              if (!res.error) {
+                this.responses = this.responses.filter(response => response.id !== id)
+              }
+            })
         }
       })
+    
   }
 
   deleteEndpoint () {  
-    this.endpointService.delete(this.endpointId)
-      .subscribe(res => {
-        if (!res.error) {
-          this.router.navigateByUrl(`/project/${database.getProject()}/folder/${this.endpoint.folder.id}`)
+    this.confirmService.open({
+      message: 'Are you sure you want to delete this endpoint'
+    })
+      .afterClose(val => {
+        if (val) {
+          this.endpointService.delete(this.endpointId)
+            .subscribe(res => {
+              if (!res.error) {
+                this.router.navigateByUrl(`/project/${database.getProject()}/folder/${this.endpoint.folder.id}`)
+              }
+            })
         }
       })
   }
